requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pycparser==2.20
python-telegram-bot==12.4.2
PyYAML==5.3.1
six==1.14.0
tornado==6.0.4
